moves blog.js into blog/index.js
